pyproject.toml
[project]

# Metadata
name = "hpcpy"
description = "A prototype Python client for interacting with HPC scheduling systems"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache 2.0 License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Operating System :: OS Independent"
]
readme = "README.md"
license = {file="LICENSE"}

# Get the version dynamically (i.e. from git)
dynamic = ["version"]

# Dependencies
requires-python = ">=3.9"
dependencies = [
    "jinja2>=3.1.4",
    "pandas>=2.2.2",
    "pytest-subprocess==1.5.3"
]

authors = [
    {name = "ACCESS-NRI", email = "access.nri@anu.edu.au"}
]

[project.urls]
source-code = "https://github.com/ACCESS-NRI/hpcpy"

[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=64",
  "versioneer[toml]==0.29",
  "jinja2>=3.1.4",
  "pandas>=2.2.2"
]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.dynamic]
version = {attr = "hpcpy.__version__"}

[tool.versioneer]
VCS = "git"
style = "pep440"
versionfile_source = "hpcpy/_version.py"
versionfile_build = "hpcpy/_version.py"
tag_prefix = ""
parentdir_prefix = "hpcpy-"